fix: repair project 043 runtime and pet boarding APIs

+++ 043-backend/src/main/java/com/xiaou/dto/PetDTO.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PetDTO {
     private Long id;
@@ -15,7 +17,7 @@ public class PetDTO {
     @NotNull(message = "年龄不能为空")
     private Integer age;
     private String gender;
-    private Double weight;
+    private BigDecimal weight;
     private String avatar;
     private String healthStatus;
     private String vaccination;
